+++ source/src/vm/mycomz80a/memory.cpp
@@ -14,7 +14,7 @@ namespace MYCOMZ80A {
 #define SET_BANK_W(s, e, w) { \
 	int sb = (s) >> 12, eb = (e) >> 12; \
 	for(int i = sb; i <= eb; i++) { \
-		if((w) == wdmy) { \
+		if(((uintptr_t)w) == ((uintptr_t)wdmy)) {	\
 			wbank[i] = wdmy; \
 		} else { \
 			wbank[i] = (w) + 0x1000 * (i - sb); \
@@ -24,7 +24,7 @@ namespace MYCOMZ80A {
 #define SET_BANK_R(s, e, r) { \
 	int sb = (s) >> 12, eb = (e) >> 12; \
 	for(int i = sb; i <= eb; i++) { \
-		if((r) == rdmy) { \
+		if(((uintptr_t)r) == ((uintptr_t)rdmy)) {	\
 			rbank[i] = rdmy; \
 		} else { \
 			rbank[i] = (r) + 0x1000 * (i - sb); \
