+++ source/src/vm/tk80bs/memory.cpp
@@ -18,12 +18,12 @@ namespace TK80 {
 #define SET_BANK(s, e, w, r) { \
 	int sb = (s) >> 9, eb = (e) >> 9; \
 	for(int i = sb; i <= eb; i++) { \
-		if((w) == wdmy) { \
+		if(((uintptr_t)w) == ((uintptr_t)wdmy)) {	\
 			wbank[i] = wdmy; \
 		} else { \
 			wbank[i] = (w) + 0x200 * (i - sb); \
 		} \
-		if((r) == rdmy) { \
+		if(((uintptr_t)r) == ((uintptr_t)rdmy)) {	\
 			rbank[i] = rdmy; \
 		} else { \
 			rbank[i] = (r) + 0x200 * (i - sb); \
